Move user and data stubs into seeded fixtures folder
