+++ cypress/fixtures/items.js
@@ -1,0 +1,36 @@
+const items = [
+  {
+    name: 'Bananas',
+    aisle: '15',
+    quantity: '2 bunches',
+    note: '',
+    inCart: false,
+    id: 1569374409096,
+  },
+  {
+    name: 'Cookies',
+    aisle: '',
+    quantity: '',
+    note: '',
+    inCart: false,
+    id: 1569370254356,
+  },
+  {
+    name: 'Ground beef',
+    aisle: '1.5',
+    quantity: '1 lb',
+    note: 'In meat section.',
+    inCart: false,
+    id: 1569374438855,
+  },
+  {
+    name: 'Apples',
+    aisle: '',
+    quantity: '2',
+    note: '',
+    inCart: false,
+    id: 1569374521481,
+  },
+];
+
+module.exports = items;
