+++ cypress/fixtures/user.js
@@ -1,0 +1,6 @@
+const user = {
+  name: 'Test',
+  password: 'password',
+};
+
+module.exports = user;
